requirements.txt
appdirs==1.4.4
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
Django==3.2.3
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.0
MarkupSafe==2.0.0
packaging==20.9
pipenv==2020.11.15
psycopg2-binary==2.8.6
PyJWT==2.1.0
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.4.6
virtualenv-clone==0.5.4
